pyproject.toml
[project]
name = "unit-convertor"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "M.Anees", email = "manees7865@gmail.com" }
]
requires-python = ">=3.12"
dependencies = [
    "litellm>=1.61.16",
    "python-decouple>=3.8",
    "python-dotenv>=1.0.1",
    "streamlit>=1.42.2",
]

[project.scripts]
unit-convertor = "unit_convertor:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
